// frontend/src/app/models/enums/Track.ts
export enum Track {
    barcelona = 'barcelona',
    brands_hatch = 'brands_hatch',
    cota = 'cota',
    donington = 'donington',
    hungaroring = 'hungaroring',
    imola = 'imola',
    indianapolis = 'indianapolis',
    kyalami = 'kyalami',
    laguna_seca = 'laguna_seca',
    misano = 'misano',
    mount_panorama = 'mount_panorama',
    monza = 'monza',
    nurburgring = 'nurburgring',
    oulton_park = 'oulton_park',
    paul_ricard = 'paul_ricard',
    silverstone = 'silverstone',
    snetterton = 'snetterton',
    spa = 'spa',
    suzuka = 'suzuka',
    watkins_glen = 'watkins_glen',
    zandvoort = 'zandvoort',
    zolder = 'zolder'
}